Separated size_of
