+++ src/ast/expr/mod.rs
@@ -4,6 +4,7 @@ mod expr;
 mod ident;
 mod literal;
 mod operator;
+mod size_of;
 mod syscall;
 mod unary;
 
@@ -13,5 +14,6 @@ pub use expr::*;
 pub use ident::*;
 pub use literal::*;
 pub use operator::*;
+pub use size_of::*;
 pub use syscall::*;
 pub use unary::*;
